requirements.txt
asgiref==3.4.1
attrs==21.2.0
blinker==1.4
certifi==2021.10.8
charset-normalizer==2.0.7
click==6.7
coreapi==2.3.3
coreschema==0.0.4
coverage==5.1
Django==3.2.8
django-cors-headers==3.3.0
django-extensions==2.2.9
django-nose==1.4.6
django-rest-swagger==2.2.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.4.0
drf-util==0.0.12
drf-yasg==1.17.1
elasticsearch==6.3.1
Faker==4.1.0
flex==6.13.2
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.2
jsonpointer==1.14
jsonschema==3.2.0
MarkupSafe==2.0.1
nose==1.3.7
openapi-codec==1.3.2
packaging==21.0
Pillow==8.3.2
PyJWT==2.2.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.3
PyYAML==3.13
requests==2.26.0
rfc3987==1.3.8
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
simplejson==3.17.5
six==1.16.0
sqlparse==0.4.2
strict-rfc3339==0.7
swagger-spec-validator==2.4.1
text-unidecode==1.3
uritemplate==4.1.1
urllib3==1.26.7
validate-email==1.3
